chore: move logic to upper level, revert: bunch of backend changes
